Add a decision controller
