+++ app/views/candidate_interface/course_choices/replace_choices/decision/choose_action.html.erb
